+++ app/Jobs/ArchiveDetteJobWithMongo.php
@@ -29,45 +29,47 @@ class ArchiveDetteJobWithMongo implements ShouldQueue
      * Execute the job.
      */
     public function handle(): void
-    {
-        // Démarrage de la transaction SQL
-        DB::beginTransaction();
+    {       
 
+        echo 'Archive debts Job started at ' . now() . ' with Mongodb';
         try {
-            $mongoClient = MongoClient::getClient();
-            $db = $mongoClient->selectDatabase(env('MONGO_DATABASE')); // Sélectionner la base de données
-            $collection = $db->selectCollection('archives_' . Carbon::now()->format('Y_m_d')); // Sélectionner la collection du jour
             
-            $dettes = Dette::with(['paiements', 'client'])->where('statut', 'solde')->get();
+            $dettes = Dette::with(['paiements', 'client', 'articles'])->where('statut', 'solde')->get();
 
-            foreach ($dettes as $dette) {
-                $data = [
-                    'dette' => $dette->toArray(),
-                    'details_dette' => $dette->getDetailsDette()->toArray(),
-                    'paiements' => $dette->paiements->toArray(),
-                    'client' => $dette->client->toArray(),
-                    'archived_at' => now(),
-                ];
+            if(!empty($dettes)){
+                DB::beginTransaction();
+                
+                $mongoClient = MongoClient::getClient();
+                $db = $mongoClient->selectDatabase('shop-universe-laravel-archive'); // Sélectionner la base de données
+                $collection = $db->selectCollection('archives_' . Carbon::now()->format('Y_m_d')); // Sélectionner la collection du jour
 
-                // Insertion dans MongoDB
-                $insertResult = $collection->insertOne($data);
-
-                // Vérification du succès de l'insertion
-                if ($insertResult->getInsertedCount() === 1) {
-                    // Supprimer les détails de la dette dans SQL
-                    DB::table('details_dette')->where('dette_id', $dette->id)->delete();
-                    $dette->paiements()->delete();
-                    $dette->delete();
-                } else {
-                    // Si l'insertion échoue, lancer une exception
-                    throw new \Exception('Échec de l\'insertion dans MongoDB');
+                foreach ($dettes as $dette) {
+                    $data = [
+                        'dette' => $dette->toArray()
+                    ];
+    
+                    // Insertion dans MongoDB
+                    $insertResult = $collection->insertOne($data);
+    
+                    // Vérification du succès de l'insertion
+                    if ($insertResult->getInsertedCount() === 1) {
+                        // Supprimer les détails de la dette dans SQL
+                        DB::table('details_dette')->where('dette_id', $dette->id)->delete();
+                        $dette->paiements()->delete();
+                        $dette->delete();
+                    } else {
+                        // Si l'insertion échoue, lancer une exception
+                        throw new \Exception('Échec de l\'insertion dans MongoDB');
+                    }
                 }
-            }
+                // Validation de la transaction SQL
+                DB::commit();
 
-            // Validation de la transaction SQL
-            DB::commit();
+                echo "Les dettes soldées ont été archivées avec succès dans MongoDB.\n";
 
-            echo "Les dettes soldées ont été archivées avec succès dans MongoDB.\n";
+            }else{
+                echo 'Aucune dette à archiver';
+            }            
 
         } catch (\Exception $e) {
             // Annuler la transaction en cas d'erreur
